+++ common/core/src/main/java/is/codion/common/state/DefaultState.java
@@ -127,7 +127,7 @@ final class DefaultState implements State {
 	}
 
 	@Override
-	public boolean isNullable() {
+	public boolean nullable() {
 		return false;
 	}