+++ common/core/src/test/java/is/codion/common/value/ValueListTest.java
@@ -40,7 +40,7 @@ public class ValueListTest {
 		assertFalse(observer.notEmpty());
 		assertUnmodifiable(observer);
 
-		assertFalse(observer.isNullable());
+		assertFalse(observer.nullable());
 		assertFalse(observer.isNull());
 		assertTrue(observer.optional().isPresent());