+++ common/core/src/test/java/is/codion/common/value/ValuesTest.java
@@ -51,7 +51,7 @@ public class ValuesTest {
 		assertFalse(observer.notEmpty());
 		assertUnmodifiable(observer);
 
-		assertFalse(observer.isNullable());
+		assertFalse(observer.nullable());
 		assertFalse(observer.isNull());
 		assertTrue(observer.optional().isPresent());
 
@@ -171,7 +171,7 @@ public class ValuesTest {
 		assertFalse(observer.notEmpty());
 		assertUnmodifiable(observer);
 
-		assertFalse(observer.isNullable());
+		assertFalse(observer.nullable());
 		assertFalse(observer.isNull());
 		assertTrue(observer.optional().isPresent());