+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/builder/AbstractComponentBuilder.java
@@ -365,7 +365,7 @@ public abstract class AbstractComponentBuilder<T, C extends JComponent, B extend
 
 	@Override
 	public final B link(Value<T> linkedValue) {
-		if (requireNonNull(linkedValue).isNullable() && !supportsNull()) {
+		if (requireNonNull(linkedValue).nullable() && !supportsNull()) {
 			throw new IllegalArgumentException("Component does not support a nullable value");
 		}
 		this.linkedValues.add(linkedValue);
@@ -374,7 +374,7 @@ public abstract class AbstractComponentBuilder<T, C extends JComponent, B extend
 
 	@Override
 	public final B link(ValueObserver<T> linkedValueObserver) {
-		if (requireNonNull(linkedValueObserver).isNullable() && !supportsNull()) {
+		if (requireNonNull(linkedValueObserver).nullable() && !supportsNull()) {
 			throw new IllegalArgumentException("Component does not support a nullable value");
 		}
 		this.linkedValueObservers.add(linkedValueObserver);