+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/NumberField.java
@@ -627,7 +627,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Short getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return isNullable() ? null : (short) 0;
+				return nullable() ? null : (short) 0;
 			}
 
 			return number.shortValue();
@@ -649,7 +649,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Double getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return isNullable() ? null : 0d;
+				return nullable() ? null : 0d;
 			}
 
 			return number.doubleValue();
@@ -671,7 +671,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Integer getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return isNullable() ? null : 0;
+				return nullable() ? null : 0;
 			}
 
 			return number.intValue();
@@ -693,7 +693,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Long getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return isNullable() ? null : 0L;
+				return nullable() ? null : 0L;
 			}
 
 			return number.longValue();
@@ -715,7 +715,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected BigDecimal getComponentValue() {
 			BigDecimal number = component().getNumber();
 			if (number == null) {
-				return isNullable() ? null : BigDecimal.ZERO;
+				return nullable() ? null : BigDecimal.ZERO;
 			}
 
 			return number;