+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - Value.BuilderFactory added, nullable() and nonNull() methods replaced with builder() returning a BuilderFactory.
 - AbstractValue.get() now implemented and final, protected getValue() added, clearValue() removed.
+- ValueObserver.isNullable() renamed back to nullable(), no factory method conflict any more.
 ### is.codion.swing.framework.model
 - SwingForeignKeyConditionModel bug fixed, IN operator only selected by default if entity is searchable.
 ### is.codion.swing.framework.ui